backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
google-generativeai
certifi
python-dotenv
pydantic